+++ packages/tools/src/tools/index.ts
@@ -15,6 +15,8 @@ import PlanarRotateTool from './PlanarRotateTool';
 import ZoomTool from './ZoomTool';
 import MIPJumpToClickTool from './MIPJumpToClickTool';
 import CrosshairsTool from './CrosshairsTool';
+import WorldCrosshairTool from './WorldCrosshairTool';
+import SliceIntersectionTool from './SliceIntersectionTool';
 import MagnifyTool from './MagnifyTool';
 import AdvancedMagnifyTool from './AdvancedMagnifyTool';
 import ReferenceLinesTool from './ReferenceLinesTool';
@@ -51,6 +53,7 @@ import KeyImageTool from './annotation/KeyImageTool';
 import AnnotationEraserTool from './AnnotationEraserTool';
 import RegionSegmentTool from './annotation/RegionSegmentTool';
 import RegionSegmentPlusTool from './annotation/RegionSegmentPlusTool';
+import ClickSegmentTool from './annotation/ClickSegmentTool';
 import WholeBodySegmentTool from './annotation/WholeBodySegmentTool';
 import LabelmapBaseTool from './segmentation/LabelmapBaseTool';
 
@@ -93,6 +96,8 @@ export {
   ReferenceCursors,
   // Annotation Tools
   CrosshairsTool,
+  WorldCrosshairTool,
+  SliceIntersectionTool,
   ReferenceLinesTool,
   OverlayGridTool,
   SegmentationIntersectionTool,
@@ -139,6 +144,7 @@ export {
   VolumeRotateTool,
   RegionSegmentTool,
   RegionSegmentPlusTool,
+  ClickSegmentTool,
   WholeBodySegmentTool,
   LabelmapBaseTool,
   SegmentBidirectionalTool,